fix delete User

+++ src/main/java/com/it/app/controller/UserController.java
@@ -5,7 +5,9 @@ import com.it.app.dto.request.UserRequestDto;
 import com.it.app.dto.response.UserResponseDto;
 import com.it.app.model.Role;
 import com.it.app.model.User;
+import com.it.app.service.InsuranceService;
 import com.it.app.service.UserService;
+import lombok.AllArgsConstructor;
 import org.dozer.Mapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,20 +21,13 @@ import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/users")
+@AllArgsConstructor
 public class UserController {
 
     private final Mapper mapper;
-
     private final UserService userService;
-
+    private final InsuranceService insuranceService;
     private final LocalizedMessageSource localizedMessageSource;
-
-
-    public UserController(Mapper mapper, UserService userService, LocalizedMessageSource localizedMessageSource) {
-        this.mapper = mapper;
-        this.userService = userService;
-        this.localizedMessageSource = localizedMessageSource;
-    }
 
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<List<UserResponseDto>> getAll() {
@@ -68,8 +63,11 @@ public class UserController {
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(value = HttpStatus.OK)
     public void delete(@PathVariable Long id) {
-        //TODO find insurance for user
-
+        //TODO find shift for user
+        //TODO delete point_user
+        if (!insuranceService.findInsurancesByUserId(id).isEmpty()) {
+            throw new RuntimeException(localizedMessageSource.getMessage("controller.user.hasInsurance", new Object[]{}));
+        }
         userService.deleteById(id);
     }
